% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid.2}
\alias{centroid.2}
\title{Convert profile data to centroid}
\usage{
centroid.2(profile, peakwidth = 0.3, angle.threshold = 0.5)
}
\arguments{
\item{profile}{A 2D spectra data matrix in profile mode}

\item{peakwidth}{The spectra peak width in \code{da} unit}
}
\value{
A 2D spectra data matrix in simple centroid mode.
}
\description{
Convert profile data to centroid
}
\details{
Convert the profiles spectrum data matrix to centroid data mode.
   The matrix data for this function shoule be a dataframe object which
   at least contains 2 column data, where:
      \code{profile <- data.frame(mz = mzlist, into = intolist);}

   If the \code{mz} or \code{into} data column is missing, then this function
   will throw an exception
}
