% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAlignment.R
\name{.alignIntensity}
\alias{.alignIntensity}
\title{Align of intensity vector}
\usage{
.alignIntensity(mz, intensity, template, mzdiff = mzkit::tolerance(0.3, "da"))
}
\arguments{
\item{mz}{the \code{m/z} vector}

\item{intensity}{the intensity vector which is corresponding to the 
\code{mz} vector.}

\item{template}{the reference \code{m/z} vector.}
}
\value{
an intensity vector which is aligned to the \code{template}
   \code{m/z} vector.
}
\description{
Align of intensity vector
}
