% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope_pattern.R
\name{isotope_pattern}
\alias{isotope_pattern}
\title{Evaluate isotope pattern}
\usage{
isotope_pattern(formula)
}
\arguments{
\item{formula}{the target formula string to evaluate}
}
\description{
Evaluate isotope pattern
}
