% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf.R
\name{mgf.ion}
\alias{mgf.ion}
\title{Generate mgf ion from data.}
\usage{
mgf.ion(mz, rt, ms2, charge = "1", title = "Unknown name ion", ms1.into = 100)
}
\arguments{
\item{ms2}{This ms2 matrix object can be in two formats:
\enumerate{
\item \code{list} ms2 data is consist with element sequence, each element must have property \code{mz} and \code{into}.
\item \code{data.frame} The data frame object should have \code{mz} and \code{into} column.
}}
}
\description{
Generate mgf ion from data.
}
