% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAlignment.R
\name{MixMS}
\alias{MixMS}
\title{Mix two MS matrix}
\usage{
MixMS(x, y, mzdiff = mzkit::tolerance(0.3, "da"))
}
\arguments{
\item{x}{a spectrum in \code{mzInto} class.}

\item{y}{a spectrum in \code{mzInto} class.}
}
\value{
a spectrum in \code{mzInto} class, which is the union
  result of two MS matrix \code{x} and \code{y}.
}
\description{
Mix two MS matrix
}
