% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MScluster.R
\name{MScluster}
\alias{MScluster}
\title{Run molecule networking}
\usage{
MScluster(peak_ms2, tolerance = mzkit::tolerance(0.3, "da"))
}
\arguments{
\item{peak_ms2}{a small bundle of MS matrix. MS matrix items
in this collection should be pre-processed by \code{centroid}}
}
\description{
Run molecule networking
}
