% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAlignment.R
\name{MScos}
\alias{MScos}
\title{SSM score in one direction}
\usage{
MScos(query, ref)
}
\arguments{
\item{query}{the MS matrix should be processed by \code{centroid}
and \code{globalAlign}}
}
\description{
SSM score in one direction
}
\details{
the MS matrix of \code{query} and \code{ref} should
    have been both pre-processed.
}
