% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAlignment.R
\name{MSDiffEntropy}
\alias{MSDiffEntropy}
\title{Evaluate the Entropy difference between two spectrum}
\usage{
MSDiffEntropy(query, ref, mzdiff = mzkit::tolerance(0.3, "da"))
}
\arguments{
\item{query}{a spectrum for query the database in \code{mzInto} class.}

\item{ref}{a reference spectrum load from the database in \code{mzInto} class.}
}
\value{
Entropy difference score between two spectrum
}
\description{
Evaluate the Entropy difference between two spectrum
}
\details{
Spectral entropy outperforms MS/MS dot
 product similarity for small-molecule compound
 identification.
}
