% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAlignment.R
\name{MSEntropy}
\alias{MSEntropy}
\title{Evaluate the Entropy of a given MS object.}
\usage{
MSEntropy(x)
}
\arguments{
\item{x}{a object with \code{mzInto} class.}
}
\description{
Evaluate the Entropy of a given MS object.
}
