% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAlignment.R
\name{MSjaccard}
\alias{MSjaccard}
\title{Jaccard index between two MS matrix}
\usage{
MSjaccard(query, ref, tolerance = mzkit::tolerance(0.3, "da"), topn = 5)
}
\arguments{
\item{query}{the MS matrix should be processed by \code{centroid}}

\item{ref}{the MS matrix should be processed by \code{centroid}}
}
\description{
Jaccard index between two MS matrix
}
\details{
the MS matrix of \code{query} and \code{ref} should
    have been both pre-processed.
}
