% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAlignment.R
\name{MStop}
\alias{MStop}
\title{Take \code{m/z} by top n intensity}
\usage{
MStop(x, topn = 5)
}
\arguments{
\item{x}{a MS matrix, and it should be pre-processed by \code{centroid}.}
}
\description{
Take \code{m/z} by top n intensity
}
