% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMgf.R
\name{parse.mgf.title_meta}
\alias{parse.mgf.title_meta}
\title{Parse meta data contains in title}
\usage{
parse.mgf.title_meta(title)
}
\arguments{
\item{title}{The title string text}
}
\description{
Parse meta data contains in title
}
