% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{precursorType.MatchImpl}
\alias{precursorType.MatchImpl}
\title{Match the precursor type}
\usage{
precursorType.MatchImpl(
  mass,
  precursorMZ,
  charge,
  chargeMode,
  tolerance,
  debug.echo
)
}
\description{
Calculate for each precursor type, and then returns the
min tolerance type as the match result
}
