% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalReader.R
\name{readAllSignals}
\alias{readAllSignals}
\title{Read of general signal pack}
\usage{
readAllSignals(cdf, verbose = TRUE)
}
\arguments{
\item{cdf}{the file path of the netCDF data file.}
}
\description{
a helper module for read sciBASIC.NET
   general signal netcdf4 data file.
}
\details{
then general signal data format can be found at source file module
   https://github.com/xieguigang/sciBASIC/blob/master/Data_science/Mathematica/SignalProcessing/Signal.IO/cdfSignalsWriter.vb
}
