% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzpackCDFReader.R
\name{readMzPackCDF}
\alias{readMzPackCDF}
\title{Read the netcdf file}
\usage{
readMzPackCDF(cdf, verbose = TRUE)
}
\arguments{
\item{cdf}{the file path of the netCDF data file.}
}
\description{
Read the netcdf file that contains the MS data converted
   from the mzPack file format.
}
\details{
the mzpack data write to netcdf file can be found at source file module:
}
