% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{toMsMatrix}
\alias{toMsMatrix}
\title{Convert \code{mzInto} object as data.frame}
\usage{
toMsMatrix(x)
}
\arguments{
\item{x}{a spectrum object in \code{mzInto} class.}
}
\value{
a dataframe with two column fields: \code{mz} and \code{into}.
}
\description{
Convert \code{mzInto} object as data.frame
}
