% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{toMzInto}
\alias{toMzInto}
\title{Convert MS dataframe as \code{mzInto} class object.}
\usage{
toMzInto(x)
}
\arguments{
\item{x}{a spectrum data object in dataframe format. the first column
in dataframe should be \code{m/z} value and the second column
in dataframe should be \code{intensity} data.}
}
\description{
Convert MS dataframe as \code{mzInto} class object.
}
