% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAlignment.R
\name{weighted_MScos}
\alias{weighted_MScos}
\title{SSM score with mass weighted}
\usage{
weighted_MScos(query, ref)
}
\description{
The fragment its m/z value is smaller,
    then weight value of this fragment is higher.
}
\details{
The \code{query} and \code{ref} spectra matrix should be aligned by
   \code{\link{globalAlign}} function at first.
}
