% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMgf.R
\name{write.mgf}
\alias{write.mgf}
\title{Write ms2 data as a mgf spectrum data file.}
\usage{
write.mgf(meta, ms2, path, verbose = TRUE)
}
\arguments{
\item{meta}{A \code{dataframe} object that contains metainfo annotation data,
this \code{dataframe} should contains at least:
\code{mz}, \code{rt}, \code{charge}, \code{libname} value fields.}

\item{ms2}{Ms2 library data. A list of \code{dataframe} object, where the list
member name is the \code{libname} value in \code{meta} parameter value.
Or this parameter can be a function to produce such data by a given libname
value.}

\item{path}{The file location of where the generated mgf file was saved.}
}
\description{
Write ms2 data as a mgf spectrum data file.
}
