% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalAlignment.R
\name{globalAlign}
\alias{globalAlign}
\title{Align \code{x} by using \code{y} as base matrix}
\usage{
globalAlign(x, y, tolerance = mzkit::tolerance(0.3, "da"))
}
\description{
Align \code{x} by using \code{y} as base matrix
}
